+++ backend/app/schemas/admin/adm_password_history.py
@@ -9,3 +9,13 @@ class AdmPasswordHistoryOut(BaseModel):
 
     class Config:
         from_attributes = True
+
+class AdmPasswordHistoryIn(BaseModel):
+    id: int
+    adm_user_id: int
+    adm_user_old_pass: str
+    created_at: None
+    updated_at: None
+
+    class Config:
+        from_attributes = True